Several changes to allow a new config option for automatically reloading groups and users...
As per a request from a user, Privileges now has an option in config.yml - `auto_reload: true`
to allow automatic reloading of the users and groups when an edit to permissions is made from
in game. This will stop the need for the /priv reload command after performing edits.

Various inspection settings have also been fixed.